Implement PartialEq, Eq, and PartialOrd on Vertex
Following:

https://www.philipdaniels.com/blog/2019/rust-equality-and-ordering/

Implemented my own based strictly on the point contained in the vertex. Two vertices should be equal if their points are equal, even if their Uuids are different (this should be an error).

Because I have f64, I cannot implement Hash or Ord without further work.

All this because I'm getting extra Facets which should be filtered out but I suspect are not because default Eq is allowing dupes.