Updates dependencies and adds `ordered-float`
Updates project dependencies to their newer versions and introduces the `ordered-float` crate. This addition enables the use of ordered floating-point numbers, which are essential for data structures requiring ordering, such as those used in geometric algorithms.

The `rand` crate is also updated with the older 0.8.5 version included as well, possibly for compatibility reasons with other dependencies.