Fix/bowyer watson (#6)
* fix: update algorithm

* feat: Add new function cell.circumsphere_contains_vertex

Supposed to be more stable, but seems to also be inaccurate in some cases.

* chore: update dependencies

* fix: CodeRabbit suggestions

* fix: removed buggy contains function

* fix: slight algorithmic improvements

* chore: update dependencies

* Updates dependencies and adds `ordered-float`

Updates project dependencies to their newer versions and introduces the `ordered-float` crate. This addition enables the use of ordered floating-point numbers, which are essential for data structures requiring ordering, such as those used in geometric algorithms.

The `rand` crate is also updated with the older 0.8.5 version included as well, possibly for compatibility reasons with other dependencies.

* Fixes hashing and comparison of floating-point points

Updates the `Point` and `Vertex` structs to use `OrderedFloat` for hashing.
This ensures that points with the same coordinates are always
hashed to the same value, regardless of their original representation.
The `Eq` derive was also removed from Vertex because it was implemented manually,
but the derive will generate a conflicting implementation.

* Improves boundary facet identification

Simplifies the check for identifying boundary facets by using `is_some_and`
instead of `map_or` for better readability and conciseness. This clarifies
the condition for determining if a cell exists and contains a given facet.

* Adds permissions to Codecov workflow

Specifies read and write permissions for contents, checks, and pull requests within the Codecov workflow.
This enhances security by granting only the necessary privileges to the workflow.