Updates Codacy analysis workflow
Updates the Codacy analysis workflow to use specific versions of the checkout and upload-sarif actions for improved stability and reliability.

Also, modifies the Codacy Analysis CLI action to use a specific commit hash and configures the project token as an input, rather than an environment variable, for better security and clarity.