Improves code analysis and CI workflows
Enhances the Codacy workflow by addressing file handling and API interactions, ensuring robust SARIF file uploads.

Adds workspace flag to rust-clippy to analyze all crates in the workspace.

Updates the spellcheck dictionary and fixes an issue in the extreme coordinates function.

Addresses a problem where the Codecov workflow was not triggered on push events to the main branch.